refactor: Rename locations.py to constants.py and consolidate DEFAULT_CLOUD_URL
- Rename openhands_cli/locations.py to openhands_cli/constants.py
- Move DEFAULT_CLOUD_URL to constants.py
- Update all imports from locations to constants
- Remove duplicate DEFAULT_CLOUD_URL definitions from main_parser.py
