+++ e2e_tests/utils.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Any
 
 from openhands.sdk import LLM
-from openhands_cli.locations import AGENT_SETTINGS_PATH, PERSISTENCE_DIR
+from openhands_cli.constants import AGENT_SETTINGS_PATH, PERSISTENCE_DIR
 from openhands_cli.utils import (
     get_default_cli_agent,
     get_llm_metadata,
