+++ openhands_cli/acp_impl/agent.py
@@ -65,7 +65,7 @@ from openhands_cli.acp_impl.utils import (
     convert_acp_mcp_servers_to_agent_format,
     convert_acp_prompt_to_message_content,
 )
-from openhands_cli.locations import CONVERSATIONS_DIR, MCP_CONFIG_FILE, WORK_DIR
+from openhands_cli.constants import CONVERSATIONS_DIR, MCP_CONFIG_FILE, WORK_DIR
 from openhands_cli.mcp.mcp_utils import MCPConfigurationError
 from openhands_cli.setup import MissingAgentSpec, load_agent_specs
 from openhands_cli.utils import extract_text_from_message_content
