+++ openhands_cli/argparsers/main_parser.py
@@ -1,7 +1,6 @@
 """Main argument parser for OpenHands CLI."""
 
 import argparse
-import os
 
 from openhands_cli import __version__
 from openhands_cli.argparsers.acp_parser import add_acp_parser
@@ -12,10 +11,7 @@ from openhands_cli.argparsers.serve_parser import add_serve_parser
 from openhands_cli.argparsers.util import add_confirmation_mode_args, add_resume_args
 from openhands_cli.argparsers.view_parser import add_view_parser
 from openhands_cli.argparsers.web_parser import add_web_parser
-
-
-# Default cloud URL
-DEFAULT_CLOUD_URL = os.getenv("OPENHANDS_CLOUD_URL", "https://app.all-hands.dev")
+from openhands_cli.constants import DEFAULT_CLOUD_URL
 
 
 def create_main_parser() -> argparse.ArgumentParser:
