+++ openhands_cli/conversations/viewer.py
@@ -11,7 +11,7 @@ from rich.console import Console
 from openhands.sdk.conversation.visualizer import DefaultConversationVisualizer
 from openhands.sdk.event.base import Event
 from openhands.tools.preset.default import register_default_tools
-from openhands_cli.locations import CONVERSATIONS_DIR
+from openhands_cli.constants import CONVERSATIONS_DIR
 from openhands_cli.theme import OPENHANDS_THEME
