+++ openhands_cli/tui/panels/mcp_side_panel.py
@@ -10,7 +10,7 @@ from textual.css.query import NoMatches
 from textual.widgets import Static
 
 from openhands.sdk import Agent
-from openhands_cli.locations import MCP_CONFIG_FILE
+from openhands_cli.constants import MCP_CONFIG_FILE
 from openhands_cli.mcp.mcp_display_utils import normalize_server_object
 from openhands_cli.mcp.mcp_utils import get_config_status
 from openhands_cli.theme import OPENHANDS_THEME
