+++ openhands_cli/tui/widgets/user_input/autocomplete_dropdown.py
@@ -4,7 +4,7 @@ from textual.containers import Container
 from textual.widgets import OptionList
 from textual.widgets.option_list import Option
 
-from openhands_cli.locations import WORK_DIR
+from openhands_cli.constants import WORK_DIR
 from openhands_cli.tui.widgets.user_input.models import (
     CompletionItem,
     CompletionType,
