+++ tests/mcp/test_mcp_commands.py
@@ -26,7 +26,7 @@ def temp_config_path():
     with tempfile.TemporaryDirectory() as temp_dir:
         config_path = Path(temp_dir) / "mcp.json"
         # Patch PERSISTENCE_DIR so that _get_mcp_config_path() returns our temp path
-        with patch("openhands_cli.locations.PERSISTENCE_DIR", str(temp_dir)):
+        with patch("openhands_cli.constants.PERSISTENCE_DIR", str(temp_dir)):
             yield config_path
