+++ tests/test_directory_separation.py
@@ -4,7 +4,7 @@ import os
 from unittest.mock import MagicMock, patch
 
 from openhands.sdk import LLM, Agent, Tool
-from openhands_cli.locations import PERSISTENCE_DIR, WORK_DIR
+from openhands_cli.constants import PERSISTENCE_DIR, WORK_DIR
 from openhands_cli.stores import AgentStore
