+++ tests/tui/panels/test_mcp_side_panel.py
@@ -218,7 +218,7 @@ class TestRefreshContentWithServerObjects:
 
         app = TestApp()
 
-        with patch("openhands_cli.locations.PERSISTENCE_DIR", str(tmp_path)):
+        with patch("openhands_cli.constants.PERSISTENCE_DIR", str(tmp_path)):
             async with app.run_test() as pilot:
                 await pilot.pause()
 
@@ -280,7 +280,7 @@ class TestRefreshContentWithServerObjects:
 
         app = TestApp()
 
-        with patch("openhands_cli.locations.PERSISTENCE_DIR", str(tmp_path)):
+        with patch("openhands_cli.constants.PERSISTENCE_DIR", str(tmp_path)):
             async with app.run_test() as pilot:
                 await pilot.pause()
 
@@ -306,7 +306,7 @@ class TestToggle:
         """Verify toggle() mounts the panel."""
         app = MCPPanelTestApp()
 
-        with patch("openhands_cli.locations.PERSISTENCE_DIR", str(tmp_path)):
+        with patch("openhands_cli.constants.PERSISTENCE_DIR", str(tmp_path)):
             async with app.run_test() as pilot:
                 await pilot.pause()
 
@@ -323,7 +323,7 @@ class TestToggle:
         """Verify toggle() removes an existing panel."""
         app = MCPPanelTestApp()
 
-        with patch("openhands_cli.locations.PERSISTENCE_DIR", str(tmp_path)):
+        with patch("openhands_cli.constants.PERSISTENCE_DIR", str(tmp_path)):
             async with app.run_test() as pilot:
                 await pilot.pause()
 
@@ -375,7 +375,7 @@ class TestToggle:
         app = MCPPanelTestApp()
 
         with (
-            patch("openhands_cli.locations.PERSISTENCE_DIR", str(tmp_path)),
+            patch("openhands_cli.constants.PERSISTENCE_DIR", str(tmp_path)),
             patch("openhands_cli.stores.AgentStore") as mock_agent_store_class,
         ):
             mock_agent_store = MagicMock()
